refactor: rename subdirectory and update imports
