+++ src/app/modules/submission/email-submission/ParsedResponsesObject.class.ts
@@ -1,5 +1,7 @@
 import { err, ok, Result } from 'neverthrow'
 
+import { formatFieldsForLogic } from 'src/shared/util/logic-utils'
+
 import {
   getLogicUnitPreventingSubmit,
   getVisibleFieldIds,
@@ -83,14 +85,17 @@ export default class ParsedResponsesObject {
     }
 
     const filteredResponses = filteredResponsesResult.value
-
+    const formattedResponses = formatFieldsForLogic(
+      filteredResponses,
+      form.form_fields,
+    )
     // Set of all visible fields
-    const visibleFieldIds = getVisibleFieldIds(filteredResponses, form)
+    const visibleFieldIds = getVisibleFieldIds(formattedResponses, form)
 
     // Guard against invalid form submissions that should have been prevented by
     // logic.
     if (
-      getLogicUnitPreventingSubmit(filteredResponses, form, visibleFieldIds)
+      getLogicUnitPreventingSubmit(formattedResponses, form, visibleFieldIds)
     ) {
       return err(new ProcessingError('Submission prevented by form logic'))
     }
