+++ src/public/modules/forms/services/form-logic/form-logic.client.service.ts
@@ -1,4 +1,4 @@
-import { isLogicCheckboxCondition } from '../../../../../shared/util/logic'
+import { isLogicCheckboxCondition } from '../../../../../shared/util/logic-utils'
 import {
   BasicField,
   IField,
