+++ src/shared/util/logic-utils/checkbox/checkbox-logic-guards.ts
