+++ src/shared/util/logic-utils/checkbox/checkbox-response-value-transformer.ts
