+++ src/shared/util/logic-utils/checkbox/index.ts
@@ -1,2 +1,2 @@
-export * from './checkbox-logic-value'
+export * from './checkbox-logic-guards'
 export * from './checkbox-response-value-transformer'
